Experiment with /f stuck
Increment by 5 instead of 1, reducing the number of passes from 250 to
50, for both X and Z. This makes the search function worst-case less
expensive, and also makes the minimum distance a player can be
teleported a minimum of 5 blocks away.

Our thoughts on what the issue here is, is that a player is teleported
to a block on the border of a claim and they’re glitched into the
block. They then move out of the glitched block (like a wall) and into
someones claim.

Under no circumstance should a player be directly teleported into
someone elses claim based on the current code.