Add a real potential fix for /f stuck
We will search all adjacent blocks for a potential /f stuck location,
disqualifying locations with a claim adjacent to it.

This will eliminate our suspected scenario where people are glitching
into claims by getting teleported into a very particular location on
the edge of a claim but in a faux safe location due to a air pocket. We
think players can then move out of the glitched block and into a base.

Worst case we add a little more processing time to /f stuck requests,
best case we fix the bug.