Add reference /kit command
Just outputs the exact same thing as the help. Many (and i encountered
a few myself) use /kit to find out info and are confused.

This command should never need to be updated as it uses the enchantment
enumerations max value, which is set somewhere else on a map per map
basis.