Add pvptimer choice / notification on /f home
If a player is in a safe zone, they will be prompted if they want to
continue teleporting home since it WILL remove their PvP timer. (Their
TP is instant)

If a player is not in a safe zone, we tell them their timer will be
removed if the teleport is not cancelled. There is a 10 second
countdown

This improves the user experience and burden on staff since they have
to explain to the player why they lost their PVP timer.