Move end portal return in overworld further out
Also set the Y back to 64, so they spawn on the road.

Furthermore, we will cache the end return location since it is static.
I would declare it final, but we initialize the location with the over
worlds reference in the portal event.

I felt this was slightly better than using Bukkit.getWorld(“world”),
even though we wont ever have any reason to not have an overworld.

This method will work perfectly fine, and avoid unnecessary object
creation which causes overhead.. even if it’s only happening on portal
events.