Fix "/resetforceinvites" not saving if no one was online.
Essentially, it was setting the value, but not flagging the team for saving, meaning the only reason it saved when people were online was because someone in the team caused it to save (edited claim, announcement, someone joined, whatever). To fix it, I simply made it flag the team for saving.