Add a 30-second "crapple" cooldown. WIP, Bukkit is buggy.
There is an issue where once the event is canceled, it will no longer
fire the consumption event for subsequent eats. In other words, if you
are holding down your right-click at 0.2s, and continue holding it down
without letting go, it will not consume your apple as the event is not
fired and was previously canceled. We will need to look into a
workaround unless I am missing something here.