Prevent scoreboard flicker for short strings.
Add scoreboard value separator indicator to prevent flicker. The flicker
seems to occur only on shorter strings. I believe this is because the
API tries to separate it on its own, and happens to put part of the time
in the score's value.

Oddly, if the separator was placed before the color code and Home and
Apple were both active on the scoreboard, the Apple timer would change
to a black color. Instead we just separate before the color.